pyproject.toml
[build-system]
  build-backend = "uv_build"
  requires = ["uv_build"]

[dependency-groups]
  dev = ["dycw-utilities[test]", "pyright", "rich"]

[project]
  authors = [{ email = "d.wan@icloud.com", name = "Derek Wan" }]
  dependencies = [
    "click >=8.3.1, <8.4",
    "dycw-utilities >=0.175.38, <0.176",
    "inflect >=7.5.0, <7.6",
    "rich >=14.2.0, <14.3",
    "ruamel-yaml >=0.19.0, <0.20",
    "tomlkit >=0.13.3, <0.14",
    "typed-settings[attrs,click] >=25.3.0, <25.4",
    "xdg-base-dirs >=6.0.2, <6.1",
  ]
  description = "Project conformalizer"
  name = "dycw-conformalize"
  readme = "README.md"
  requires-python = ">= 3.12"
  version = "0.11.13"

  [project.scripts]
    conformalize = "conformalize.cli:_main"

[tool]
  [tool.uv]
    [tool.uv.build-backend]
      module-name = "conformalize"
      module-root = "src"
